requirements.txt
# ClaimFlow AI - Requirements

# Core Framework
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Data Visualization
plotly>=5.18.0
pandas>=2.1.0

# AI & LLM Integration
google-genai>=1.0.0
python-dotenv>=1.0.0

# RAG & Vector Database
sentence-transformers>=2.0.0
chromadb>=0.4.0

# Voice Features (Optional)
SpeechRecognition>=3.10.0
gTTS>=2.5.0
# pydub

# Document Processing
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Language Detection
langdetect>=1.0.9

# Development
# pytest
# black
# flake8
